+++ disk_space_other.go
@@ -1,0 +1,13 @@
+//go:build !windows
+
+package main
+
+import "syscall"
+
+func availableDiskBytes(directory string) (uint64, error) {
+	var stat syscall.Statfs_t
+	if err := syscall.Statfs(directory, &stat); err != nil {
+		return 0, err
+	}
+	return stat.Bavail * uint64(stat.Bsize), nil
+}